# internal/models/user.go
package models

import "time"

type Role string

const (
	RoleUser      Role = "user"
	RolePowerUser Role = "power_user"
	RoleAdmin     Role = "admin"
)

func (r Role) Valid() bool {
	switch r {
	case RoleUser, RolePowerUser, RoleAdmin:
		return true
	}
	return false
}

type User struct {
	ID           int64     `json:"id"`
	Username     string    `json:"username"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"-"`
	Role         Role      `json:"role"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
}